requirements.txt
streamlit
python-dotenv
requests
pandas
pygame
